+++ TotalColumn.php
@@ -4,6 +4,7 @@ namespace bupy7\grid;
 
 use Closure;
 use yii\helpers\ArrayHelper;
+use yii\base\InvalidConfigException;
 
 /**
  * Total column. Adding to last row total formula for all items from this column.
@@ -54,7 +55,18 @@ class TotalColumn extends DataColumn
      * ~~~
      */
     public $footer = self::FORMULA_SUM;
-    
+    /**
+     * @var string|array|Closure Format of total value in a footer. It can be string either array for Formatter
+     * component or an anonymous function where the first argument is calculated value.
+     * ~~~
+     * function ($value) {
+     *      return Yii::t('app', 'UNIT_N', $value);
+     * }
+     * ~~~
+     * @since 1.1.5
+     */
+    public $footerFormat ='text';
+
     private $_data = [];   
     private $_models = [];
     
@@ -91,12 +103,13 @@ class TotalColumn extends DataColumn
     protected function renderFooterCellContent()
     {
         $footer = $this->calculateSummary();
-        return trim($footer) !== '' ? $footer : $this->grid->emptyCell;
+        return trim($footer) !== '' ? $this->formattingFooter($footer) : $this->grid->emptyCell;
     }
     
     /**
      * Calculates the summary of an input data based on page summary aggregration function.
      * @return mixed
+     * @throws InvalidConfigException
      */
     protected function calculateSummary()
     {
@@ -129,10 +142,28 @@ class TotalColumn extends DataColumn
                     break;
 
                 default:
-                    $result = null;
+                    throw new InvalidConfigException('There is invalid value `$footer`.');
             }
         }
         return $result;
-    }  
+    }
+
+    /**
+     * Formatting footer value uses Formatter component or an anonymous function.
+     * @param string $value
+     * @return mixed
+     * @since 1.1.5
+     * @throws InvalidConfigException
+     */
+    protected function formattingFooter($value)
+    {
+        if (is_string($this->footerFormat) || is_array($this->footerFormat)) {
+            return $this->grid->formatter->format($value, $this->footerFormat);
+        } elseif ($this->footerFormat instanceof Closure) {
+            return call_user_func($this->footerFormat, $value);
+        } else {
+            throw new InvalidConfigException('There is invalid value `$footerFormat`.');
+        }
+    }
 }